+++ PIPS-IPM/Core/QpGen/QpGenVars.C
@@ -73,7 +73,7 @@ QpGenVars::QpGenVars( OoqpVector * x_in, OoqpVector * s_in,
 }
 
 QpGenVars::QpGenVars( LinearAlgebraPackage * la,
-		      int nx_, int my_, int mz_,
+		      long long nx_, long long my_, long long mz_,
 		      OoqpVector * ixlow_in, OoqpVector * ixupp_in,
 		      OoqpVector * iclow_in, OoqpVector * icupp_in )
 {
